pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "werkzeug>=3.1.3",
]
